+++ src/Constants/CustomHeaders.php
@@ -12,4 +12,5 @@ class CustomHeaders
     public const IDEMPOTENCY_KEY = 'Idempotency-Key';
     public const CORRELATION_ID = 'X-Tl-Correlation-Id';
     public const WEBHOOK_TIMESTAMP = 'X-Tl-Webhook-Timestamp';
+    public const TL_AGENT = 'TL-Agent';
 }