+++ src/Entities/Payment/PaymentRetrieved/PaymentSettled.php
@@ -5,12 +5,23 @@ declare(strict_types=1);
 namespace TrueLayer\Entities\Payment\PaymentRetrieved;
 
 use DateTimeInterface;
+use TrueLayer\Exceptions\ApiRequestJsonSerializationException;
+use TrueLayer\Exceptions\ApiResponseUnsuccessfulException;
+use TrueLayer\Exceptions\InvalidArgumentException;
+use TrueLayer\Exceptions\SignerException;
+use TrueLayer\Exceptions\ValidationException;
+use TrueLayer\Interfaces\HasApiFactoryInterface;
 use TrueLayer\Interfaces\Payment\PaymentSettledInterface;
 use TrueLayer\Interfaces\Payment\PaymentSourceInterface;
+use TrueLayer\Interfaces\Payment\RefundRequestInterface;
+use TrueLayer\Interfaces\Payment\RefundRetrievedInterface;
+use TrueLayer\Traits\ProvidesApiFactory;
 use TrueLayer\Validation\ValidType;
 
-final class PaymentSettled extends _PaymentWithAuthorizationConfig implements PaymentSettledInterface
+final class PaymentSettled extends _PaymentWithAuthorizationConfig implements PaymentSettledInterface, HasApiFactoryInterface
 {
+    use ProvidesApiFactory;
+
     /**
      * @var PaymentSourceInterface
      */
@@ -84,5 +95,51 @@ final class PaymentSettled extends _PaymentWithAuthorizationConfig implements Pa
     public function getSettledAt(): DateTimeInterface
     {
         return $this->settledAt;
+    }
+
+    /**
+     * @return RefundRequestInterface
+     * @throws ValidationException
+     *
+     * @throws InvalidArgumentException
+     */
+    public function refund(): RefundRequestInterface
+    {
+        return $this->make(RefundRequestInterface::class)
+            ->payment($this->getId());
+    }
+
+    /**
+     * @param string $refundId
+     *
+     * @return RefundRetrievedInterface
+     * @throws ApiResponseUnsuccessfulException
+     * @throws InvalidArgumentException
+     * @throws SignerException
+     * @throws ValidationException
+     *
+     * @throws ApiRequestJsonSerializationException
+     */
+    public function getRefund(string $refundId): RefundRetrievedInterface
+    {
+        $data = $this->getApiFactory()->paymentsApi()->retrieveRefund($this->getId(), $refundId);
+
+        return $this->make(RefundRetrievedInterface::class, $data);
+    }
+
+    /**
+     * @return RefundRetrievedInterface[]
+     * @throws ApiResponseUnsuccessfulException
+     * @throws InvalidArgumentException
+     * @throws SignerException
+     * @throws ValidationException
+     *
+     * @throws ApiRequestJsonSerializationException
+     */
+    public function getRefunds(): array
+    {
+        $data = $this->getApiFactory()->paymentsApi()->retrieveRefunds($this->getId());
+
+        return $this->makeMany(RefundRetrievedInterface::class, $data);
     }
 }