+++ src/Entities/Payment/Refund/RefundPending.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TrueLayer\Entities\Payment\Refund;
+
+use TrueLayer\Interfaces\Payment\RefundPendingInterface;
+
+class RefundPending extends RefundRetrieved implements RefundPendingInterface
+{
+}