+++ src/Factories/ClientFactory.php
@@ -94,6 +94,7 @@ final class ClientFactory implements ClientFactoryInterface
 
         $authClient = new ApiClient($this->httpClient, $authBaseUri);
         $authClient = new Decorators\ExponentialBackoffDecorator($authClient);
+        $authClient = new Decorators\TLAgentDecorator($authClient);
 
         $this->authToken = new AccessToken(
             $authClient,
@@ -134,6 +135,7 @@ final class ClientFactory implements ClientFactoryInterface
         $this->apiClient = new Decorators\ExponentialBackoffDecorator($this->apiClient);
         $this->apiClient = new Decorators\SigningDecorator($this->apiClient, $signer);
         $this->apiClient = new Decorators\IdempotencyKeyDecorator($this->apiClient);
+        $this->apiClient = new Decorators\TLAgentDecorator($this->apiClient);
     }
 
     /**